+++ package/rhythmicon-vue/index.js
@@ -1,6 +1,7 @@
 import RhythmSequencer from "./components/RhythmSequencer.vue"
 import RhythmInput from "./components/RhythmInput.vue"
+import RhythmControls from "./components/RhythmControls.vue"
 import RhythmCircle from "./components/RhythmCircle.vue"
 import RhythmScore from "./components/RhythmScore.vue"
 
-export { RhythmSequencer, RhythmInput, RhythmCircle, RhythmScore }
+export { RhythmSequencer, RhythmInput, RhythmControls, RhythmCircle, RhythmScore }